Add files via upload
Summarize the project and what problem it was solving.
This project focused on assisting with a banking app. The app was created to show the initial investment amount, monthly deposit, annual interest, and number of years.  It was challenging overall to learn how to solve each project. I tried coding everything individually but everything connecting together was difficult. many times the code would have errors. 
What did you do particularly well?
Attempting to code and trying to make it all work together was something that even though I did not do as well on I tried. I am hopeful that with more practice it will help. 
Where could you enhance your code? How would these improvements make your code more efficient, secure, and so on? I would dedicate more time to code each section. This class takes time and dedication to solve. It is not easy. 
Which pieces of the code did you find most challenging to write, and how did you overcome this? What tools or resources are you adding to your support network? Overall it is not the lines but fixing mistakes that have been done throughout. learning how to debug a code has been the most challenging. I would attempt to rely more on ly resources to help me move forward. 
What skills from this project will be particularly transferable to other projects or coursework? How to begin a code to ensure the names you are selecting make sense. this will be crucial when learning how to fix it. 
How did you make this program maintainable, readable, and adaptable? I had to go back many times on some sections. re-read and try to understand the concepts.